feat: block and delete friend done

+++ django/src/friend/views.py
@@ -19,23 +19,33 @@ class UserRelationViewSet(viewsets.ModelViewSet):
     serializer_class = UserRelationSerializer
     
     def get_queryset(self):
-        return UserRelation.objects.filter(user=self.request.user)
+        return UserRelation.objects.filter(user=self.request.user, deleted=False)
 
-    @action(detail=True, methods=['post'])
-    def block(self, request, pk=None):
-        user_relation = get_object_or_404(UserRelation, pk=pk)
+    @action(detail=False, methods=['post'])
+    def block(self, request):
+        friend_id = request.data.get('friend_id')
+        if not friend_id:
+            return Response({'error': 'friend_id is required'}, status=400)
+        user_relation = get_object_or_404(UserRelation, user=request.user, friend=friend_id)
         user_relation.block()
         return Response({'status': 'friend blocked'})
 
-    @action(detail=True, methods=['post'])
-    def unblock(self, request, pk=None):
-        user_relation = get_object_or_404(UserRelation, pk=pk)
+    @action(detail=False, methods=['post'])
+    def unblock(self, request):
+        friend_id = request.data.get('friend_id')
+        if not friend_id:
+            return Response({'error': 'friend_id is required'}, status=400)
+        user_relation = get_object_or_404(UserRelation, user=request.user, friend=friend_id)
         user_relation.unblock()
         return Response({'status': 'friend unblocked'})
 
-    @action(detail=True, methods=['delete'])
-    def delete(self, request, pk=None):
-        user_relation = get_object_or_404(UserRelation, pk=pk)
+    @action(detail=False, methods=['delete'])
+    def delete(self, request):
+        friend_id = request.query_params.get('friend_id')
+        if not friend_id:
+            return Response({'error': 'friend_id is required'}, status=status.HTTP_400_BAD_REQUEST)
+        
+        user_relation = get_object_or_404(UserRelation, user=request.user, friend=friend_id)
         user_relation.delete()
         return Response({'status': 'friend deleted'}, status=status.HTTP_204_NO_CONTENT)
     
@@ -47,13 +57,14 @@ class UserRelationViewSet(viewsets.ModelViewSet):
         user = get_object_or_404(User, username=username)
         if user == request.user:
             return Response({'status': 'cannot add yourself as a friend'}, status=status.HTTP_400_BAD_REQUEST)
-        elif UserRelation.objects.filter(user=request.user, friend=user).exists():
+        elif UserRelation.objects.filter(user=request.user, friend=user, deleted=False).exists():
             return Response({'status': 'already friends'}, status=status.HTTP_400_BAD_REQUEST)
         profile = get_object_or_404(Profile, user=user)
         serializer = ProfileSerializer(profile)
         response_data = serializer.data
-        can_add_friend = not FriendRequest.objects.filter(sender=request.user, receiver=user, status=FriendRequest.Status.ACCEPTED).order_by('-created_at').first()
-        response_data['is_friend'] = user.friends.filter(id=request.user.id).exists()
+        is_friend = user.friends.filter(id=request.user.id, deleted=False).exists()
+        can_add_friend = not is_friend and FriendRequest.objects.filter(sender=request.user, receiver=user).order_by('-created_at').first().status != FriendRequest.Status.PENDING
+        response_data['is_friend'] = is_friend
         response_data['can_add_friend'] = can_add_friend
         return Response(response_data, status=status.HTTP_200_OK)
